+++ BE/src/main/java/com/ssafy/hangbokdog/sponsorship/domain/repository/SponsorshipHistoryJdbcRepository.java
@@ -26,7 +26,7 @@ public class SponsorshipHistoryJdbcRepository {
 			public void setValues(PreparedStatement ps, int i) throws SQLException {
 				SponsorshipHistory sponsorshipHistory = sponsorshipHistories.get(i);
 				ps.setLong(1, sponsorshipHistory.getSponsorshipId());
-				ps.setLong(2, sponsorshipHistory.getAmount());
+				ps.setInt(2, sponsorshipHistory.getAmount());
 				ps.setString(3, sponsorshipHistory.getStatus().toString());
 			}
