+++ BE/src/main/java/com/ssafy/hangbokdog/sponsorship/dto/ActiveSponsorshipInfo.java
@@ -8,6 +8,6 @@ public record ActiveSponsorshipInfo(
 	String dogName,
 	Long mileageId,
 	Long balance,
-	Long amount
+	int amount
 ) {
 }
